# BACKEND/pic/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: PIC DIAN
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of the responses to the client (Accepts)
produces:
  - application/json
  - application/xml
paths:
...
  /agregarPregunta:
      x-swagger-router-controller: preguntas_w_controller
      post:
        description: agregar una pregunta
        operationId: agregarPregunta
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /traerPreguntas:
      x-swagger-router-controller: preguntas_w_controller
      post:
        description: obtener preguntas
        operationId: obtenerPreguntasNuevas
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /agregarCuestionario:
      x-swagger-router-controller: cuestionario.controller
      post:
        description: agregar cuestionario
        operationId: Cuestionario
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /buscarPreguntasPorID:
      x-swagger-router-controller: preguntas_w_controller
      post:
        description: buscar pregunta por id
        operationId: buscarPreguntasPorID
        responses:
          "200":
            description: Success
            schema:  
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /preguntasPorIDCuestionario:
      x-swagger-router-controller: preguntas_w_controller
      post:
        description: buscar pregunta por id para el cuestionaio
        operationId: buscarPreguntasPorIDCuestionario
        responses:
          "200":
            description: Success
            schema:  
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /buscarCompetenciasCuestionario:
      x-swagger-router-controller: preguntas_w_controller
      post:
        description: Actualizar pregunta iii
        operationId: buscarCompetenciasCuestionario
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /limpiarPreguntas:
      x-swagger-router-controller: preguntas_w_controller
      post: 
        description: limpia la info de las preguntas
        operationId: limpiarPreguntas
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
definitions:
  GenericResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string